news-ai-project/requirements.txt
crewai
langchain_groq
streamlit
python-dotenv
requests
beautifulsoup4
lxml